requirements.txt
flask==2.0.3
numpy==1.19.5
torch==1.10.0
transformers==4.15.0
opencv-python-headless==4.5.5.64
pywin32==300
python-dotenv==0.19.2
gunicorn==20.1.0
datasets
eventlet==0.33.3
flask_socketio==5.3.6
psutil==5.9.0
pymodbus==2.5.3
py2neo==2021.2.4
ultralytics==8.0.20
redis==4.5.4
pydantic==1.10.7
fastapi==0.95.0
uvicorn==0.21.1
neo4j==5.5.0
torchvision==0.11.1
torchaudio==0.10.0
scikit-learn==1.0.2
pillow==9.5.0
requests==2.28.1
beautifulsoup4==4.12.2
lxml==4.9.2
python-multipart==0.0.5
librosa==0.10.1
langdetect==1.0.9
pyserial==3.5
